Bringing Back Styleguide CSS and attempting to fix nav bar
Brought back the styleguide css because the pages Styleguide, Citations, and Wireframes would not work without it. I renamed it from main.css to styleguide.css. Also attempted readjust the nav bar, but was only able to update the fonts to be the correct families. Also added a little down arrow to the drowdown elements so it is clear to users that these are dropdown.